do not overwrite app.ini if it exists in persistent storage

+++ templates/init/_container.tpl
@@ -12,10 +12,12 @@ Create helm partial for gitea server
         name: {{ template "postgresql.fullname" . }}
         key: postgres-password
   - name: SCRIPT
-    value: &script |- 
+    value: &script |-
       mkdir -p /datatmp/gitea/conf
-      cp -f /etc/gitea/app.ini /datatmp/gitea/conf/app.ini
-      sed -i "s/POSTGRES_PASSWORD/${POSTGRES_PASSWORD}/g" /datatmp/gitea/conf/app.ini
+      if [ ! -f /datatmp/gitea/conf/app.ini ]; then
+        cp -f /etc/gitea/app.ini /datatmp/gitea/conf/app.ini
+        sed -i "s/POSTGRES_PASSWORD/${POSTGRES_PASSWORD}/g" /datatmp/gitea/conf/app.ini
+      fi
   command: ["/bin/sh",'-c', *script]
   volumeMounts:
   - name: gitea-data
